moved Upload to the end of stepper, bugs
