+++ app/build.gradle
@@ -108,7 +108,7 @@ dependencies {
 
     implementation 'com.github.GoodieBag:PinView:1.4'
 
-    implementation 'com.theartofdev.edmodo:android-image-cropper:2.8.+'
+    implementation 'com.theartofdev.edmodo:android-image-cropper:2.8.0'
 }
 
 configurations.all {
