+++ app/src/main/java/com/example/meetingapp/fragments/profile_stepper/UserBasicInformationStepperFragmentFragment.java
@@ -10,6 +10,7 @@ import android.content.pm.PackageManager;
 import android.database.Cursor;
 import android.graphics.Bitmap;
 import android.net.Uri;
+import android.os.AsyncTask;
 import android.os.Bundle;
 import android.provider.MediaStore;
 import android.util.Log;
@@ -17,6 +18,7 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.DatePicker;
+import android.widget.FrameLayout;
 import android.widget.RadioButton;
 import android.widget.RadioGroup;
 import android.widget.RelativeLayout;
@@ -58,6 +60,8 @@ import java.util.Locale;
 import java.util.Map;
 import java.util.Objects;
 
+import javax.xml.transform.Result;
+
 import butterknife.BindView;
 import butterknife.ButterKnife;
 import butterknife.OnClick;
@@ -71,7 +75,7 @@ import retrofit2.Response;
 import static android.app.Activity.RESULT_OK;
 
 public class UserBasicInformationStepperFragmentFragment extends Fragment implements BlockingStep,
-        DatePickerDialog.OnDateSetListener, GetImageFromAsync {
+        DatePickerDialog.OnDateSetListener {
 
     private static final int REQUEST_EXTERNAL_STORAGE = 1;
     private static String[] PERMISSIONS_STORAGE = {
@@ -90,14 +94,33 @@ public class UserBasicInformationStepperFragmentFragment extends Fragment implem
     TextView headerH4;
     private IUserProfileManager iUserProfileManager;
     private FragmentActivity mContext;
-    private String imageUrl;
     private Bitmap bitmap;
     private String pattern = "yyyy-MM-dd";
     private Date date;
     private String sex;
 
     private Uri mCropImageUri;
-    private String currentPhotoPath;
+
+    private class CompressBitmap extends AsyncTask<Bitmap, Integer, byte[]>{
+
+        @Override
+        protected byte[] doInBackground(Bitmap... bitmaps) {
+            ByteArrayOutputStream stream = new ByteArrayOutputStream();
+            bitmaps[0].compress(Bitmap.CompressFormat.PNG, 100, stream);
+            return stream.toByteArray();
+        }
+
+        @Override
+        protected void onPreExecute() {
+            super.onPreExecute();
+        }
+
+        @Override
+        protected void onPostExecute(byte[] bytes) {
+            super.onPostExecute(bytes);
+            iUserProfileManager.savePhoto(bytes);
+        }
+    }
 
     public static UserBasicInformationStepperFragmentFragment newInstance() {
         return new UserBasicInformationStepperFragmentFragment();
@@ -158,20 +181,6 @@ public class UserBasicInformationStepperFragmentFragment extends Fragment implem
         CropImage.startPickImageActivity(requireContext(), this);
     }
 
-    private String getRealPathFromURI(Uri contentURI) {
-        String result;
-        Cursor cursor = getContext().getContentResolver().query(contentURI, null, null, null, null);
-        if (cursor == null) { // Source is Dropbox or other similar local file path
-            result = contentURI.getPath();
-        } else {
-            cursor.moveToFirst();
-            int idx = cursor.getColumnIndex(MediaStore.Images.ImageColumns.DATA);
-            result = cursor.getString(idx);
-            cursor.close();
-        }
-        return result;
-    }
-
     public void onActivityResult(int requestCode, int resultCode, Intent data) {
         if (requestCode == CropImage.PICK_IMAGE_CHOOSER_REQUEST_CODE && resultCode == Activity.RESULT_OK) {
             Uri imageUri = CropImage.getPickImageResultUri(this.requireContext(), data);
@@ -216,13 +225,20 @@ public class UserBasicInformationStepperFragmentFragment extends Fragment implem
         String date = Objects.requireNonNull(textBirthDate.getText()).toString();
         iUserProfileManager.saveBirthDate(date);
         iUserProfileManager.saveSex(sex);
-        upload(mCropImageUri);
+        iUserProfileManager.saveUri(mCropImageUri);
 
-        ByteArrayOutputStream stream = new ByteArrayOutputStream();
-        bitmap.compress(Bitmap.CompressFormat.PNG, 100, stream);
-        byte[] byteArray = stream.toByteArray();
+        layoutAvatarMask.setVisibility(View.GONE);
+        imageProfile.setImageBitmap(bitmap);
+        //upload(mCropImageUri);
 
-        iUserProfileManager.savePhoto(byteArray);
+//        ByteArrayOutputStream stream = new ByteArrayOutputStream();
+//        bitmap.compress(Bitmap.CompressFormat.PNG, 100, stream);
+//        byte[] byteArray = stream.toByteArray();
+//
+//        iUserProfileManager.savePhoto(byteArray);
+
+        CompressBitmap compressBitmap = new CompressBitmap();
+        compressBitmap.execute(bitmap);
 
         callback.goToNextStep();
     }
@@ -313,56 +329,10 @@ public class UserBasicInformationStepperFragmentFragment extends Fragment implem
         }
     }
 
-    private void upload(Uri imageUri) {
-
-        String fullPath = getRealPathFromURI(imageUri);
-        File file = new File(fullPath);
-
-        File  compressFile = Compressor.getDefault(getContext()).compressToFile(file);
-
-        RequestBody requestFile = RequestBody.create(compressFile, MediaType.parse(fullPath));
-
-        Map<String, RequestBody> map = new HashMap<>();
-        map.put("photo\"; filename=\"" + fullPath + "\"", requestFile);
-
-
-        RetrofitClient.needsHeader(true);
-        RetrofitClient.setToken(PreferenceUtils.getToken(requireContext()));
-        Call<ProfilePhoto> call = RetrofitClient
-                .getInstance(PreferenceUtils.getToken(requireContext()))
-                .getApi()
-                .uploadFile(map);
-
-        call.enqueue(new Callback<ProfilePhoto>() {
-            @Override
-            public void onResponse(@NonNull Call<ProfilePhoto> call, @NonNull Response<ProfilePhoto> response) {
-                ProfilePhoto profilePhoto = response.body();
-                if (profilePhoto != null) {
-                    Toast.makeText(getContext(), "Loaded!", Toast.LENGTH_SHORT).show();
-
-                    imageUrl = profilePhoto.getPhoto();
-                    new DownloadImageTask(UserBasicInformationStepperFragmentFragment.this).execute(imageUrl);
-                }
-            }
-
-            @Override
-            public void onFailure(@NonNull Call<ProfilePhoto> call, @NonNull Throwable t) {
-                Log.d("error", Objects.requireNonNull(t.getMessage()));
-            }
-        });
-    }
-
     @Override
     public void onResume() {
         super.onResume();
         if (bitmap != null)
             imageProfile.setImageBitmap(bitmap);
-    }
-
-    @Override
-    public void getResult(Bitmap bitmap) {
-        layoutAvatarMask.setVisibility(View.GONE);
-        imageProfile.setImageBitmap(bitmap);
-        this.bitmap = bitmap;
     }
 }
