+++ app/src/main/java/com/example/meetingapp/fragments/profile_stepper/UserPublishStepperFragment.java
@@ -4,44 +4,57 @@ import android.annotation.SuppressLint;
 import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
+import android.database.Cursor;
 import android.graphics.Bitmap;
 import android.graphics.BitmapFactory;
+import android.net.Uri;
 import android.os.Bundle;
+import android.provider.MediaStore;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.fragment.app.Fragment;
 
+import com.example.meetingapp.interfaces.GetImageFromAsync;
 import com.example.meetingapp.interfaces.IUserProfileManager;
 import com.example.meetingapp.R;
 import com.example.meetingapp.activities.MainActivity;
 import com.example.meetingapp.api.RetrofitClient;
 import com.example.meetingapp.models.Category;
+import com.example.meetingapp.models.ProfilePhoto;
 import com.example.meetingapp.models.UserProfile;
 import com.example.meetingapp.utils.PreferenceUtils;
+import com.example.meetingapp.utils.images.DownloadImageTask;
+import com.example.meetingapp.utils.images.compression.Compressor;
 import com.google.android.material.chip.Chip;
 import com.google.android.material.chip.ChipGroup;
 import com.stepstone.stepper.BlockingStep;
 import com.stepstone.stepper.StepperLayout;
 import com.stepstone.stepper.VerificationError;
 
+import java.io.File;
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Objects;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
 import de.hdodenhof.circleimageview.CircleImageView;
+import okhttp3.MediaType;
+import okhttp3.RequestBody;
 import retrofit2.Call;
 import retrofit2.Callback;
 import retrofit2.Response;
 
-public class UserPublishStepperFragment extends Fragment implements BlockingStep {
+public class UserPublishStepperFragment extends Fragment implements BlockingStep, GetImageFromAsync {
 
 
     @BindView(R.id.text_birth_date)
@@ -107,6 +120,9 @@ public class UserPublishStepperFragment extends Fragment implements BlockingStep
 
     @Override
     public void onCompleteClicked(StepperLayout.OnCompleteClickedCallback callback) {
+        if (iUserProfileManager.getUri()!=null){
+            upload(iUserProfileManager.getUri());
+        } else Log.d("@@@@@@@@@@@@@@@@@@@@@@", "onCompleteClicked: no uri");
         createUserProfile();
     }
 
@@ -201,5 +217,64 @@ public class UserPublishStepperFragment extends Fragment implements BlockingStep
     @Override
     public void onError(@NonNull VerificationError error) {
 
+    }
+
+    private void upload(Uri imageUri) {
+
+        String fullPath = getRealPathFromURI(imageUri);
+        File file = new File(fullPath);
+
+        File  compressFile = Compressor.getDefault(getContext()).compressToFile(file);
+
+        RequestBody requestFile = RequestBody.create(compressFile, MediaType.parse(fullPath));
+
+        Map<String, RequestBody> map = new HashMap<>();
+        map.put("photo\"; filename=\"" + fullPath + "\"", requestFile);
+
+
+        RetrofitClient.needsHeader(true);
+        RetrofitClient.setToken(PreferenceUtils.getToken(requireContext()));
+        Call<ProfilePhoto> call = RetrofitClient
+                .getInstance(PreferenceUtils.getToken(requireContext()))
+                .getApi()
+                .uploadFile(map);
+
+        call.enqueue(new Callback<ProfilePhoto>() {
+            @Override
+            public void onResponse(@NonNull Call<ProfilePhoto> call, @NonNull Response<ProfilePhoto> response) {
+                ProfilePhoto profilePhoto = response.body();
+                if (profilePhoto != null) {
+                    Toast.makeText(getContext(), "Loaded!", Toast.LENGTH_SHORT).show();
+
+                    String imageUrl = profilePhoto.getPhoto();
+                    new DownloadImageTask(UserPublishStepperFragment.this).execute(imageUrl);
+                }
+            }
+
+            @Override
+            public void onFailure(@NonNull Call<ProfilePhoto> call, @NonNull Throwable t) {
+                Log.d("error", Objects.requireNonNull(t.getMessage()));
+            }
+        });
+    }
+
+    private String getRealPathFromURI(Uri contentURI) {
+        String result;
+        Cursor cursor = getContext().getContentResolver().query(contentURI, null, null, null, null);
+        if (cursor == null) { // Source is Dropbox or other similar local file path
+            result = contentURI.getPath();
+        } else {
+            cursor.moveToFirst();
+            int idx = cursor.getColumnIndex(MediaStore.Images.ImageColumns.DATA);
+            result = cursor.getString(idx);
+            cursor.close();
+        }
+        return result;
+    }
+
+    @Override
+    public void getResult(Bitmap bitmap) {
+        imageProfile.setImageBitmap(bitmap);
+        this.bitmap = bitmap;
     }
 }
